+++ expressions/Integer.h
@@ -6,7 +6,7 @@ class Integer : public Expression
 {
 public:
     Integer(const int value);
-    double evaluate() const;
+    int evaluate() const;
 
 private:
     const int m_value;
