+++ expressions/operators/Division.h
@@ -6,5 +6,5 @@ class Division : public BinaryOperator
 {
 public:
     Division(ExpressionPtr left, ExpressionPtr right);
-    double evaluate() const;
+    int evaluate() const;
 };
