+++ expressions/operators/Subtraction.cpp
@@ -5,7 +5,7 @@ Subtraction::Subtraction(ExpressionPtr left, ExpressionPtr right)
 {
 }
 
-double Subtraction::evaluate() const
+int Subtraction::evaluate() const
 {
     return m_left->evaluate() - m_right->evaluate();
 }
